Individual file update 5 (optimization of command based, some clean up)
This is what I have done in the last 11 days regarding the code. The code is now makes the practice bot drive in the right direction. Some lines were removed as they were not needed. I implemented an autonomous command in the command based robot. Encoders are initialized in the command based robot but are not operational yet.